feat: take half a day of unpaid leave, morning or afternoon
A working day can now be half unpaid leave. It costs half a day of basic and
none of the allowance -- the same rule as a whole day off, halved -- because a
half day stays in the basic's divisor and carries half its weight in the
allowance's.

There are two kinds because the money is not the only thing on screen. The
ticker follows the working window, and a single "half day" would have had to
guess which half was worked: guessing the morning counts all morning for
someone at the dentist and stops at noon just as they arrive. So an afternoon
off keeps the start of the day and a morning off keeps its end, each half the
paid time long and, like a Saturday half day, not stopping for lunch. On a day
that is already a half day, the halves split that day and anchor to its own end,
not to clock-off.

One number carries all of it. DayOverride.workedFraction feeds payWeight,
paidSeconds and the allowance divisor, so the clock and the wallet cannot
disagree about a half day. isWorkday had meant "no override", under which the
half that is worked would never have ticked, and now means "any of it worked".

Overtime is refused on an afternoon off. It is counted from the end of the
window, which on an afternoon off is one o'clock, so without that rule the hours
taken off would have been paid as overtime. A morning off can still run late.

The click cycle is unchanged; the halves live in a right-click menu on the Mac
and a touch-and-hold menu on the iPhone, which offers every state with a tick on
the current one. In the cycle they would have turned the commonest correction,
unpaid back to a workday, from one click into three. The grid draws a half day
as the two halves it is, and the days-off count reads 1.5 rather than calling an
afternoon a day.

Twenty tests. Thirteen deliberate bugs were planted, each against the test meant
to catch it, and all thirteen were caught; each catching test was then confirmed
to run and pass on clean code, so no catch came from a filter matching nothing.
Verified on the simulator against figures worked out beforehand: $131.87 at
16:50 on a morning off and $137.16 after an afternoon off, both to the cent.

An adversarial review found four wording faults, now fixed -- the French legend
reused the word for a scheduled half day, and the Korean one meant paid annual
leave -- and a crash in the debug renderer. That crash predates this change:
on the 30th of every month the two existing preview examples clamped onto the
same day, and a duplicate key in a dictionary literal traps. Checked across
every day of 2026 and 2027: the old code crashed on 226 of 730, the new on none.

Two older bugs it found are left for their own changes: the watch complication
freezes during overtime on any day, and goal ready-times do not clip lunch when
lunch falls outside working hours.

+++ Sources/Domain/Entities/Earnings.swift
@@ -41,7 +41,7 @@ public struct Earnings: Equatable, Sendable {
     /// Working days of this month already finished, today excluded.
     public var workdaysCompletedThisMonth: Int
     /// Scheduled days this month marked as holiday or leave.
-    public var daysOffThisMonth: Int
+    public var daysOffThisMonth: Double
     /// Paid seconds worked so far today.
     public var elapsedPaidSeconds: TimeInterval
     public var totalPaidSeconds: TimeInterval
@@ -59,7 +59,7 @@ public struct Earnings: Equatable, Sendable {
         monthEarned: Double,
         workdaysThisMonth: Int,
         workdaysCompletedThisMonth: Int,
-        daysOffThisMonth: Int = 0,
+        daysOffThisMonth: Double = 0,
         elapsedPaidSeconds: TimeInterval,
         totalPaidSeconds: TimeInterval,
         progress: Double,
